HBX-1653: Move class 'org.hibernate.tool.hbm2x.ConfigurationNavigator' to 'org.hibernate.tool.internal.export.common.ConfigurationNavigator'
