+++ orm/src/main/java/org/hibernate/tool/hbm2x/GenericExporter.java
@@ -13,6 +13,7 @@ import org.hibernate.mapping.Component;
 import org.hibernate.tool.hbm2x.pojo.ComponentPOJOClass;
 import org.hibernate.tool.hbm2x.pojo.POJOClass;
 import org.hibernate.tool.internal.export.common.AbstractExporter;
+import org.hibernate.tool.internal.export.common.ConfigurationNavigator;
 
 
 public class GenericExporter extends AbstractExporter {
